+++ wlts/datasources/ds_manager.py
@@ -10,20 +10,27 @@ from json import loads as json_loads
 
 import pkg_resources
 
-from .wfs import WFSDataSource
 from .wcs import WCSDataSource
+from .wfs import WFSDataSource
 
 
 class DataSourceFactory:
-    """Factory for DataSource."""
+    """Factory Class for DataSource."""
 
     @staticmethod
     def make(ds_type, id, conn_info):
-        """Factory method for creates datasource.
+        """Factory method for creates a datasource.
 
-        New datasources must be add in factorys.
-        Ex: factorys = {"POSTGIS": "PostGisDataSource", "WCS": "WCSDataSource", "WFS": "WFSDataSource",
-                        "RASTER FILE": "RasterFileDataSource"}
+        Args:
+            ds_type (str): The datasource type to be create.
+            id (str): The datasource identifier.
+            conn_info (dict): The datasource connection information.
+
+        .. note::
+            New datasources must be add in factorys.
+
+            Ex: factorys = {"POSTGIS": "PostGisDataSource", "WCS": "WCSDataSource", \
+                            "WFS": "WFSDataSource", "RASTER FILE": "RasterFileDataSource"}
         """
         factorys = {"WFS": "WFSDataSource", "WCS": "WCSDataSource"}
         datasource = eval(factorys[ds_type])(id, conn_info)
@@ -31,7 +38,7 @@ class DataSourceFactory:
 
 
 class DataSourceManager:
-    """DataSourceManager Class."""
+    """This is a singleton to manage all datasource instances available."""
 
     _datasources = list()
 
@@ -53,7 +60,17 @@ class DataSourceManager:
         return DataSourceManager.__instance
 
     def get_datasource(self, ds_id):
-        """Get DataSource."""
+        """Return a datasource object.
+
+        Args:
+            ds_id (str): Identifier of a datasource.
+
+        Returns:
+            datasource: A datasource available in the server.
+
+        Raises:
+            RuntimeError: If the datasource not found.
+        """
         try:
             for ds in self._datasources:
                 if ds.get_id == ds_id:
@@ -62,11 +79,15 @@ class DataSourceManager:
             raise RuntimeError(f"Datasource identifier {ds_id} not found in WLTS Datasources!")
 
     def insert_datasource(self, conn_info):
-        """Insert DataSource."""
+        """Creates a new datasource and stores in list of datasource.
+
+        Args:
+            conn_info (dict): The datasource connection information.
+        """
         self._datasources.append(DataSourceFactory.make(conn_info["type"], conn_info["id"], conn_info))
 
     def load_all(self):
-        """Load All DataSources."""
+        """Creates all datasource based on json of datasource."""
         json_string = pkg_resources.resource_string('wlts', '/json_configs/datasources.json').decode('utf-8')
         config = json_loads(json_string)