+++ wlts/ext.py
@@ -12,6 +12,7 @@ from flask_sqlalchemy import SQLAlchemy
 
 db = SQLAlchemy()
 
+
 class WLTSDatabase:
     """WLTS Database."""